cv_uploader/cv-upload-handler/requirements.txt
boto3==1.9.4
python-docx==0.8.4
